+++ src/ErrorHandler.php
@@ -69,7 +69,7 @@ final class ErrorHandler
         $renderer ??= $this->defaultRenderer;
 
         try {
-            $this->logger->error(PlainTextRenderer::throwableToString($t), ['throwable' => $t]);
+            $this->logger->error($t->getMessage(), ['throwable' => $t]);
             return $this->debug ? $renderer->renderVerbose($t, $request) : $renderer->render($t, $request);
         } catch (Throwable $t) {
             return new ErrorData((string) $t);